Use available sounds assets for sound options
